requirements.txt
fastapi
uvicorn
httpx
python-dotenv
playwright
pydantic
